# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Proyectos",
        "contact": {
            "email": "andy.ancajima@mimco.com.pe"
        },
        "version": "1"
    },
    "paths": {
...
        "/GestionProyectos/public/index.php/vali_etap": {
            "post": {
                "tags": [
                    "Gestion etapa"
                ],
                "summary": "obtiene datos estapa por el idetapa",
                "operationId": "App\\Http\\Controllers\\EtapaController::vali_etap",
                "parameters": [
                    {
                        "name": "intIdEtapa",
                        "in": "path",
                        "description": "Seleccionamos el idetapa que vamos obtener la informacion",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdEtapa": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdEtapa": "5"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sin Mensaje"
                    },
                    "407": {
                        "description": "El Documento de identidad ingresado no se encuentra registrado."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
...
        "/GestionProyectos/public/index.php/regi_tipo_etap": {
            "post": {
                "tags": [
                    "Gestion Tipo Etapa"
                ],
                "summary": "permite registrar el tipo etapa",
                "operationId": "App\\Http\\Controllers\\TipoEtapasController::regi_tipo_etap",
                "parameters": [
                    {
                        "name": "intIdAgru",
                        "in": "path",
                        "description": "Selecciona un IdAgrupador",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "6"
                    },
                    {
                        "name": "varCodiTipoEtap",
                        "in": "path",
                        "description": "Ingresamos el codigo del tipo etapa que se va registrar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "EMPA_PRUEBA"
                    },
                    {
                        "name": "varDescTipoEtap",
                        "in": "path",
                        "description": "Descripcion del nuevo tipo de etapa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "EMPAQUE_PRUEBA"
                    },
                    {
                        "name": "varEstaTipoEtap",
                        "in": "path",
                        "description": "estado",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ACT"
                    },
                    {
                        "name": "acti_usua",
                        "in": "path",
                        "description": "Usuario que va registrar el nuevo tipo etapa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "montañez_andy"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdTipoEtap": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdAgru": "6",
                                    "varCodiTipoEtap": "EMPA_PRUEBA",
                                    "varDescTipoEtap": "EMPAQUE_PRUEBA",
                                    "varEstaTipoEtap": "ACTIVO",
                                    "acti_usua": "montañez_andy"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro Satisfactorio"
                    },
                    "407": {
                        "description": "Código de Tipo Etapa Ya existe."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
...
        "/GestionProyectos/public/index.php/vali_tipo_etap": {
            "post": {
                "tags": [
                    "Gestion Tipo Etapa"
                ],
                "summary": "obtiene datos del tipo etapa que se ha seleccionado",
                "operationId": "App\\Http\\Controllers\\TipoEtapasController::vali_tipo_etap",
                "parameters": [
                    {
                        "name": "intIdTipoEtap",
                        "in": "path",
                        "description": "El idTipoEtap que se va obtener su informacion",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdTipoEtap": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdTipoEtap": "5"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sin Mensaje"
                    },
                    "407": {
                        "description": "El idTipoEtapa no se encuentra registrado."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
        "/GestionProyectos/public/index.php/actu_tipo_etap": {
            "post": {
                "tags": [
                    "Gestion Tipo Etapa"
                ],
                "summary": "Permite actualizar el tipo etapa.",
                "operationId": "App\\Http\\Controllers\\TipoEtapasController::actu_tipo_etap",
                "parameters": [
                    {
                        "name": "intIdTipoEtap",
                        "in": "path",
                        "description": "El idTipoEtap que se modificara",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "120"
                    },
                    {
                        "name": "intIdAgru",
                        "in": "path",
                        "description": "Selecciona un IdAgrupador",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "120"
                    },
                    {
                        "name": "varCodiTipoEtap",
                        "in": "path",
                        "description": "Ingresamos el codigo del tipo etapa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "EMPA_PRUEBA_MODI"
                    },
                    {
                        "name": "varDescTipoEtap",
                        "in": "path",
                        "description": "Descripcion del nuevo tipo de etapa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "EMPAQUE_PRUEBA_MODI"
                    },
                    {
                        "name": "varEstaTipoEtap",
                        "in": "path",
                        "description": "Estado de tipo etapa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ACT"
                    },
                    {
                        "name": "usua_modi",
                        "in": "path",
                        "description": "Usuario que va modificar el tipo etapa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "montañez_andy"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdTipoEtap": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdTipoEtap": "120",
                                    "intIdAgru": "120",
                                    "varCodiTipoEtap": "EMPA_PRUEBA_MODI",
                                    "varDescTipoEtap": "EMPAQUE_PRUEBA_MODI",
                                    "varEstaTipoEtap": "ACT",
                                    "usua_modi": "usuario"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizacion Satisfactorio"
                    },
                    "407": {
                        "description": "Código de Tipo Etapa Ya existe."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
        "/GestionProyectos/public/index.php/list_tipo_etapa_segun_agru": {
            "post": {
                "tags": [
                    "Gestion Tipo Etapa"
                ],
                "summary": "obtener el agrupador mediante el id de tipo etapa",
                "operationId": "App\\Http\\Controllers\\TipoEtapasController::list_tipo_etapa_segun_agru",
                "parameters": [
                    {
                        "name": "intIdTipoEtap",
                        "in": "path",
                        "description": "Ingrese el id Tipo etapa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdTipoEtap": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdTipoEtap": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Muestra el Agrupador"
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
...
    }
}